+++ src/net.rs
@@ -1,34 +1,24 @@
 use std::{
     collections::{HashMap, VecDeque},
     io::{self, ErrorKind},
-    net::{SocketAddr, SocketAddrV4, SocketAddrV6, UdpSocket},
+    net::{SocketAddr, UdpSocket},
     os::unix::io::{AsRawFd, RawFd},
     sync::atomic::{AtomicBool, Ordering}
 };
 
 use super::util::{MockTimeSource, Time, TimeSource};
 
-use net2::UdpBuilder;
-
 
 pub trait Socket: AsRawFd + Sized {
-    fn listen_v4(host: &str, port: u16) -> Result<Self, io::Error>;
-    fn listen_v6(host: &str, port: u16) -> Result<Self, io::Error>;
+    fn listen(addr: SocketAddr) -> Result<Self, io::Error>;
     fn receive(&mut self, buffer: &mut [u8]) -> Result<(usize, SocketAddr), io::Error>;
     fn send(&mut self, data: &[u8], addr: SocketAddr) -> Result<usize, io::Error>;
     fn address(&self) -> Result<SocketAddr, io::Error>;
 }
 
 impl Socket for UdpSocket {
-    fn listen_v4(host: &str, port: u16) -> Result<Self, io::Error> {
-        UdpBuilder::new_v4().expect("Failed to obtain ipv4 socket builder").bind((host, port))
-    }
-    fn listen_v6(host: &str, port: u16) -> Result<Self, io::Error> {
-        UdpBuilder::new_v6()
-            .expect("Failed to obtain ipv4 socket builder")
-            .only_v6(true)
-            .expect("Failed to set only_v6")
-            .bind((host, port))
+    fn listen(addr: SocketAddr) -> Result<Self, io::Error> {
+        UdpSocket::bind(addr)
     }
     fn receive(&mut self, buffer: &mut [u8]) -> Result<(usize, SocketAddr), io::Error> {
         self.recv_from(buffer)
@@ -99,13 +89,8 @@ impl AsRawFd for MockSocket {
 }
 
 impl Socket for MockSocket {
-    fn listen_v4(host: &str, port: u16) -> Result<Self, io::Error> {
-        let ip = try_fail!(host.parse(), "Failed to parse IPv4 address: {}");
-        Ok(Self::new(SocketAddr::V4(SocketAddrV4::new(ip, port))))
-    }
-    fn listen_v6(host: &str, port: u16) -> Result<Self, io::Error> {
-        let ip = try_fail!(host.parse(), "Failed to parse IPv6 address: {}");
-        Ok(Self::new(SocketAddr::V6(SocketAddrV6::new(ip, port, 0, 0))))
+    fn listen(addr: SocketAddr) -> Result<Self, io::Error> {
+        Ok(Self::new(addr))
     }
     fn receive(&mut self, buffer: &mut [u8]) -> Result<(usize, SocketAddr), io::Error> {
         if let Some((addr, data)) = self.inbound.pop_front() {