+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [semantic versioning](http://semver.org).
 
+### UNRELEASED
+
+- [added] Added option to listen on specified IP
+- [changed] No longer using two sockets for ipv4 and ipv6
+
 ### v1.3.0 (2020-01-25)
 
 - [added] Building for aarch64 aka arm64 (thanks to Ivan)